+++ src/PjbServer/Tools/Console/CommandRepository.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace PjbServer\Tools\Console;
+
+class CommandRepository
+{
+
+    /**
+     * @var array
+     */
+    protected $commands;
+
+    public function __construct()
+    {
+        $this->commands = [
+            'pjbserver:start' => new Command\PjbServerStartCommand(),
+            'pjbserver:restart' => new Command\PjbServerRestartCommand(),
+            'pjbserver:stop' => new Command\PjbServerStopCommand()
+        ];
+    }
+
+    /**
+     * @param $name
+     * @return \Symfony\Component\Console\Command\Command
+     */
+    public function getRegisteredCommand($name)
+    {
+        return $this->commands[$name];
+    }
+
+    /**
+     * Return all registered commands
+     * @return array
+     */
+    public function getRegisteredCommands()
+    {
+        return $this->commands;
+    }
+}